test: working with dynamic components to populate default values
